feat: move models to shared (course mentor recomendation)
