+++ youtube_client_app/src/app/youtube/components/detailed-card/detailed-card.component.html
@@ -1,0 +1,1 @@
+<p>detailed-card works!</p>
