+++ youtube_client_app/src/app/youtube/components/detailed-card/detailed-card.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-detailed-card',
+  templateUrl: './detailed-card.component.html',
+  styleUrls: ['./detailed-card.component.scss'],
+})
+export class DetailedCardComponent {}
